# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "BookMyEvent API",
    "description": "Ticketing system API for event organisers to create & manage events.\nUsers can purchase/reserve tickets for events.\n",
    "version": "1.0.0",
    "contact": {
      "name": "API Support",
      "email": "support@bookmyevent.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
...
  "paths": {
...
    "/venues/{id}/seats/bulk": {
      "post": {
        "operationId": "VenuesController_createBulkSeats",
        "summary": "Create bulk seats for a venue",
        "tags": ["Venues"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" },
            "example": "550e8400-e29b-41d4-a716-446655440000"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VenueCreateBulkSeatDto"
              },
              "example": {
                "rows": ["A", "B", "C"],
                "startNumber": 1,
                "endNumber": 50
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Seats created successfully",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "data": [
                    {
                      "id": "aa0e8400-e29b-41d4-a716-446655440000",
                      "venueId": "550e8400-e29b-41d4-a716-446655440000",
                      "row": "A",
                      "number": 1
                    },
                    {
                      "id": "aa0e8400-e29b-41d4-a716-446655440001",
                      "venueId": "550e8400-e29b-41d4-a716-446655440000",
                      "row": "A",
                      "number": 2
                    }
                  ],
                  "message": "150 seats created successfully"
                }
              }
            }
          }
        }
      }
    },
...
    "/seats": {
      "post": {
        "operationId": "SeatsController_create",
        "summary": "Create a new seat",
        "tags": ["Seats"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateSeatDto" },
              "example": {
                "eventId": "660e8400-e29b-41d4-a716-446655440000",
                "row": "A",
                "number": 1,
                "ticketId": "770e8400-e29b-41d4-a716-446655440000"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Seat created successfully",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "data": {
                    "id": "990e8400-e29b-41d4-a716-446655440000",
                    "eventId": "660e8400-e29b-41d4-a716-446655440000",
                    "seatId": "SEAT-001",
                    "status": "AVAILABLE",
                    "createdAt": "2026-04-18T12:00:00.000Z"
                  },
                  "message": "Seat created successfully"
                }
              }
            }
          }
        }
      },
      "post/bulk": {
        "operationId": "SeatsController_createBulk",
        "summary": "Create multiple seats",
        "tags": ["Seats"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateBulkSeatDto" },
              "example": {
                "eventId": "660e8400-e29b-41d4-a716-446655440000",
                "rows": ["A", "B", "C"],
                "startNumber": 1,
                "endNumber": 50,
                "ticketId": "770e8400-e29b-41d4-a716-446655440000"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Seats created successfully",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "data": [
                    {
                      "id": "990e8400-e29b-41d4-a716-446655440000",
                      "eventId": "660e8400-e29b-41d4-a716-446655440000",
                      "seatId": "SEAT-001",
                      "status": "AVAILABLE"
                    },
                    {
                      "id": "990e8400-e29b-41d4-a716-446655440001",
                      "eventId": "660e8400-e29b-41d4-a716-446655440000",
                      "seatId": "SEAT-002",
                      "status": "AVAILABLE"
                    }
                  ],
                  "message": "5 seats created successfully"
                }
              }
            }
          }
        }
      },
      "post/lock": {
        "operationId": "SeatsController_lockSeats",
        "summary": "Lock seats for booking",
        "tags": ["Seats"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LockSeatsDto" },
              "example": {
                "seatIds": [
                  "990e8400-e29b-41d4-a716-446655440000",
                  "990e8400-e29b-41d4-a716-446655440001"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Seats locked",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "data": [
                    {
                      "id": "990e8400-e29b-41d4-a716-446655440000",
                      "status": "LOCKED"
                    },
                    {
                      "id": "990e8400-e29b-41d4-a716-446655440001",
                      "status": "LOCKED"
                    }
                  ],
                  "message": "Seats locked"
                }
              }
            }
          }
        }
      },
      "post/book": {
        "operationId": "SeatsController_bookSeats",
        "summary": "Book seats",
        "tags": ["Seats"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/BookSeatsDto" },
              "example": {
                "seatIds": [
                  "990e8400-e29b-41d4-a716-446655440000",
                  "990e8400-e29b-41d4-a716-446655440001"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Seats booked successfully",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "data": [
                    {
                      "id": "990e8400-e29b-41d4-a716-446655440000",
                      "status": "BOOKED"
                    },
                    {
                      "id": "990e8400-e29b-41d4-a716-446655440001",
                      "status": "BOOKED"
                    }
                  ],
                  "message": "Seats booked successfully"
                }
              }
            }
          }
        }
      },
      "post/release": {
        "operationId": "SeatsController_releaseSeats",
        "summary": "Release seats",
        "tags": ["Seats"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ReleaseSeatsDto" },
              "example": {
                "seatIds": [
                  "990e8400-e29b-41d4-a716-446655440000",
                  "990e8400-e29b-41d4-a716-446655440001"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Seats released successfully",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "data": [
                    {
                      "id": "990e8400-e29b-41d4-a716-446655440000",
                      "status": "AVAILABLE"
                    },
                    {
                      "id": "990e8400-e29b-41d4-a716-446655440001",
                      "status": "AVAILABLE"
                    }
                  ],
                  "message": "Seats released successfully"
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "SeatsController_findAll",
        "summary": "Get all seats",
        "tags": ["Seats"],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 0 }
          },
          {
            "name": "take",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 50 }
          },
          {
            "name": "sortBy",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "default": "createdAt" }
          },
          {
            "name": "sortOrder",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["ASC", "DESC"],
              "default": "DESC"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": { "type": "string" }
          },
          {
            "name": "eventId",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "format": "uuid" },
            "example": "660e8400-e29b-41d4-a716-446655440000"
          }
        ],
        "responses": {
          "200": {
            "description": "List of seats",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "data": [
                    {
                      "id": "990e8400-e29b-41d4-a716-446655440000",
                      "eventId": "660e8400-e29b-41d4-a716-446655440000",
                      "seatId": "SEAT-001",
                      "status": "AVAILABLE",
                      "createdAt": "2026-04-18T12:00:00.000Z"
                    }
                  ],
                  "pagination": { "skip": 0, "take": 10, "total": 1 }
                }
              }
            }
          }
        }
      }
    },
    "/seats/{id}": {
      "get": {
        "operationId": "SeatsController_findOne",
        "summary": "Get a seat by ID",
        "tags": ["Seats"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" },
            "example": "990e8400-e29b-41d4-a716-446655440000"
          }
        ],
        "responses": {
          "200": {
            "description": "Seat found",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "data": {
                    "id": "990e8400-e29b-41d4-a716-446655440000",
                    "eventId": "660e8400-e29b-41d4-a716-446655440000",
                    "seatId": "SEAT-001",
                    "status": "AVAILABLE",
                    "createdAt": "2026-04-18T12:00:00.000Z"
                  }
                }
              }
            }
          },
          "404": { "description": "Seat not found" }
        }
      },
      "patch": {
        "operationId": "SeatsController_update",
        "summary": "Update a seat",
        "tags": ["Seats"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" },
            "example": "990e8400-e29b-41d4-a716-446655440000"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateSeatDto" },
              "example": { "status": "BOOKED" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Seat updated successfully",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "data": {
                    "id": "990e8400-e29b-41d4-a716-446655440000",
                    "status": "BOOKED"
                  },
                  "message": "Seat updated successfully"
                }
              }
            }
          },
          "404": { "description": "Seat not found" }
        }
      },
      "delete": {
        "operationId": "SeatsController_remove",
        "summary": "Delete a seat",
        "tags": ["Seats"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" },
            "example": "990e8400-e29b-41d4-a716-446655440000"
          }
        ],
        "responses": {
          "204": { "description": "Seat deleted successfully" },
          "404": { "description": "Seat not found" }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "VenueCreateBulkSeatDto": {
        "type": "object",
        "properties": {
          "rows": {
            "type": "array",
            "items": { "type": "string" },
            "description": "Array of row identifiers (e.g., ['A', 'B', 'C'])"
          },
          "startNumber": {
            "type": "integer",
            "description": "Start seat number"
          },
          "endNumber": { "type": "integer", "description": "End seat number" }
        },
        "required": ["rows", "startNumber", "endNumber"]
      },
...
      "CreateSeatDto": {
        "type": "object",
        "properties": {
          "eventId": {
            "type": "string",
            "format": "uuid",
            "description": "Event ID"
          },
          "row": {
            "type": "string",
            "description": "Seat row (e.g., 'A', 'B')"
          },
          "number": { "type": "integer", "description": "Seat number" },
          "ticketId": {
            "type": "string",
            "format": "uuid",
            "description": "Ticket type ID"
          }
        },
        "required": ["eventId", "row", "number"]
      },
      "CreateBulkSeatDto": {
        "type": "object",
        "properties": {
          "eventId": {
            "type": "string",
            "format": "uuid",
            "description": "Event ID"
          },
          "rows": {
            "type": "array",
            "items": { "type": "string" },
            "description": "Array of row identifiers (e.g., ['A', 'B', 'C'])"
          },
          "startNumber": {
            "type": "integer",
            "description": "Start seat number"
          },
          "endNumber": { "type": "integer", "description": "End seat number" },
          "ticketId": {
            "type": "string",
            "format": "uuid",
            "description": "Ticket type ID"
          }
        },
        "required": ["eventId", "startNumber", "endNumber"]
      },
      "UpdateSeatDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": ["AVAILABLE", "LOCKED", "BOOKED"]
          }
        }
      },
      "LockSeatsDto": {
        "type": "object",
        "properties": {
          "seatIds": {
            "type": "array",
            "items": { "type": "string", "format": "uuid" },
            "description": "Array of seat IDs to lock"
          }
        },
        "required": ["seatIds"]
      },
      "BookSeatsDto": {
        "type": "object",
        "properties": {
          "seatIds": {
            "type": "array",
            "items": { "type": "string", "format": "uuid" },
            "description": "Array of seat IDs to book"
          }
        },
        "required": ["seatIds"]
      },
      "ReleaseSeatsDto": {
        "type": "object",
        "properties": {
          "seatIds": {
            "type": "array",
            "items": { "type": "string", "format": "uuid" },
            "description": "Array of seat IDs to release"
          }
        },
        "required": ["seatIds"]
      },
...
    }
  }
}